refactor to pull TestData out as shared component
